Fix erlang diff formatter for file creation
By checking for '--- a/' we were missing the files that have just been
created as they instead start with '--- /dev/null'.
Instead we should look for the new file name with '+++ b/'.
This should also fix the case when a file is renamed, which we likely
wouldn't have handled correctly previously.

Change-Id: Ie518753c9a8aba692ad7a4fad665551d6f0c306f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193856
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>