MB-57867: Fix erlc build flag: bin_opt_info
This flag is used by erlc to determine situations where binaries are
being used inefficiently. We already had what appeared to be this option
in our rebar config but it was mispelled (words mixed up order).

It was orginally: bin_info_opt
The flag was actually: bin_opt_info

I came across this when wondering if I was working with a binary in the
most efficient way.. It seemed as though we wanted to have this enabled,
all along, but we have not.

Unfortunately, if we were to properly enable it, we would run into
warnings that would fail the build. We should probably look into these
warnings, at the very least to make sure we are working with binaries
efficiently.

So instead of including the option, I am going to completely remove it
but create a jira ticket to track running this manually by setting the
environment variable and then trackign down the issues.

  $ export ERL_COMPILER_OPTIONS=bin_opt_info
  $ make

This will give us the warnings and we can work on them from there. This
is a technical dept issue we probably meant to be paying attention to
all along.

Change-Id: I7ba176feee81dbdf770d38b7288ddccc17fef598
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193872
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>