MB-55578: Simplify kv_stats_monitor refresh mechanism
We store a reference to the timer so that we can delay a refresh , but
there's no good reason to do this. Simplify the code so that we can rely
on the health monitor behaviour code to do this for us in the future.

Change-Id: I4babdcf06d8052cc5d4dbf2ba3b5aa40957a948a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194881
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>