MB-57924 Pass maxTTL=0 to memcached to disable TTL
This fixes the issue where specifying maxTTL=0 during the creation of a
collection or when patching a collection doesn't pass on the maxTTL in
the manifest to memcached. As a result the bucket's maxTTL is used.

The fix leads to maxTTL=0 getting passed to memcached in the manifest.

* create collection without specifying maxTTL
  [{name,<<"mycol3">>},{uid,<<"c">>}]
* create collection with maxTTL=0 (new behavior)
  [{name,<<"mycol3">>},{maxTTL,0},{uid,<<"c">>}]
* create collection with maxTTL=123
  [{name,<<"mycol3">>},{maxTTL,123},{uid,<<"c">>}]

This also affects what is returned via the REST interface. Before this
change we would return maxTTL=0 even if the user hadn't specified it.

          "name": "mycol1",
          "uid": "a",
          "maxTTL": 0,
          "history": false

With this change maxTTL isn't returned if it wasn't specified.

          "name": "mycol1",
          "uid": "a",
          "history": false

Change-Id: Ice77dfd04e6cd5663c463a1a419ee2556261ce2f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194540
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>