MB-57924 Provide way to reset collection maxTTL
Prior to this change there was no way to "reset" the maxTTL of a
collection to a value which specified "use the maxTTL of the bucket, if
one has been specified".

This change provides 'maxTTL=bucket' which...
* on collection creation is a no-op as the default is to not have a
  collection specific maxTTL
* on collection modification, using -X PATCH, resets the maxTTL for the
  collection to "use the bucket maxTTL"

As an example:
 /pools/default/buckets/<bucket>/scopes/<scope>/collections \
    -d 'name=mycollection'

creates a collection named mycollection with no maxTTL specified. The
maxTTL value of the bucket will be used.

 /pools/default/buckets/<bucket>/scopes/<scope>/collections/mycollection \
    -d 'maxTTL=123' -X PATCH

modifies the collection named mycollection to have a maxTTL value of
123. This overrides any bucket maxTTL value.

 /pools/default/buckets/<bucket>/scopes/<scope>/collections/mycollection \
    -d 'maxTTL=bucket' -X PATCH

modifies the collect named mycollection to not have a maxTTL. The maxTTL
value of the bucket will be used.

Change-Id: I19523bfe6919035ff1d62bf05a541b70300b3f56
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194613
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>