[dialyzer] Use proplist type in place of pattern
This patch is to replace all the uses of "[{_,_}]" with
proplists:proplist() type specification in the ns_bucket file where it's
been used to approximate a proplist. Our original pattern was actually
more narrowly defined than the proplists:proplist() type spec. The only
difference being that proplists:proplist() also allows empty atom-keyed
entries. The number of elements (size) of the tuple is also potentially
different. The specification is actually:

-type property()  :: atom() | tuple().
-type proplist()  :: [property()].

Which means this is a valid proplist: [{key1, "value1"}, key2, key3] and
the keys key2 and key3 are also valid, but just have the value of
"true" (Ex: from functions like proplists:lookup/2).

Change-Id: I47630a0a2ed8bafd2bcf5e18ebeaccd7876fb151
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194834
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>