MB-48217: [SM] Make sure password can be reset to ""...
... when default settings are used

This is needed for backward compatibility

Note: if gosecrets is configured to use the password command,
the password param in

POST /node/controller/changeMasterPassword

must not be set.
In that case this endpoint triggers the password command execution
and returned value is used as a new password.

Change-Id: I8c138028388edb047cb397d51296d04b5ba8670b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194917
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>