[cluster_tests]: Test travel-sample loading on ...
... on two node cluster.

In MB-58093, MB-58009 it was seen that travel-sample loading was broken,
when cbimport used cbauth. Surprisingly it works on a one node cluster,
but fails on a two node cluster.

I presume it's because we pass down the special password to cbimport (on
the node it is spawned), but in a two node cluster the special password
for the second node is fetched from cbauth and that kept failing.

Moving the requirements of this testset to perform the sample loading on
a two node cluster.

Change-Id: If405a245637453873230af5b6a69213395e6ed0c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195006
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Hareen Kancharla <hareen.kancharla@couchbase.com>