MB-57924 Use -1 to reset collection's maxTTL
Per PM the recently introduced "maxTTL=bucket" to reset a collection's
maxTTL value to "use the bucket's maxTTL if the bucket has one" has been
changed to "maxTTL=-1".

maxTTL=-1 is used by PUT for
/pools/default/buckets/<bucket>/scopes/<scope>/collections to change
the maxTTL for a collection to "use the bucket's maxTTL".

It is also used by backup/restore to restore a manifest where the
collection's maxTTL needs to be reset to "use the bucket's maxTTL".

maxTTL=-1 for a collection is never in a manifest created by ns_server
nor passed in a manifest to kv.

Change-Id: I2e0c39937eac5b6353a5d1688dd679df75d05686
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195222
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>