MB-55578, MB-56973: Set dcp_proxy liveliness_update_interval to 100ms
With a recent change KV should send NOOP messages every 100ms if no
other messages are send in that time frame. KV should do this if the two
connected nodes are running 7.6.0 or newer. To faster detect failures we
should also update the liveliness update interval in the dcp_proxy
accordingly to 100ms. There are some interesting scenarios worth
exploring here around cross-version clusters. Should the cluster be
running in mixed mode then the dcp_traffic_monitor will run at most
every 1 second. It will use the last time that it heard from the
dcp_proxy which is the last time that the dcp_proxy processed a liveness
check and had seen a new message since the previous check. This,
interestingly, allows us to update the liveness check interval without
trigger unnecessary auto-failovers in mixed mode as the dcp traffic
monitor should expect to see the same number of messages. It is only
when the dcp_traffic_monitor refresh interval is scaled below 1s
(when auto-failover timeout is below 5s) that we need more frequent
liveness checks as inactivity is measured in multiples of the tick
interval.

In addition, make this parameter configurable.

Change-Id: I0fda7da04d41c02caf6b07aa8943101528d814b5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195734
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>