MB-57260: Use bucket priority for failover
Use the newly created bucket "priority" setting to determine the order
in which buckets should be failed-over. The higher the number the
greater the priority, and the earlier in the process the bucket will be
failed-over. Defaults to 0 for buckets in which it is not explicitly
set. Max priority is equal to the maximum number of buckets that are
allowed. That means it can change depending on the deployment or using
various overrides.

Change-Id: Ic7e25fdec62ab53de1ca3d19494d64331dd15ab0
Reviewed-on: https://review.couchbase.org/c/ns_server/+/194157
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>