MB-55578: Allow index_monitor to scale refresh interval below 1s
For the sake of faster-failover. This is a different change to the
behaviour change so that it can go in a different build.

This change scales the refresh interval when the auto-failover timeout
value is configured below 5 seconds. When configured below 5 seconds the
refresh interval of the monitor is scaled to 1/10th of that value.

Change-Id: Ia4b85dccc713bba24ed45ebea7be722093a9482a
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195514
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: <varun.velamuri@couchbase.com>