MB-24487: [SAML] Redirect user back to UI if authz fails
If SAML authentication succeeds, user gets redirected to
/saml/consume which is supposed to validate the assertion.
If assertion is valid (user is authenticted) it checks
if that user has permission to use the UI.
If so, it sets the UI cookie and redirects the user back to UI.
Otherwise we should also redirect the user to the UI, but
1. Do not set the cookie
2. Show proper error, so the user can understand what went wrong

In that error it is important to show the username that was
extracted from the SAML assertion, because it should help
understanding what went wrong (is that incorrect name extracted,
or is that missing external couchbase user).

The problem is we can't pass username and other sensitive
information in url to the UI when redirecting (because this would be
a leak of PI in logs), so we store it in ns_server instead, and pass
ID of that information in URL to the UI. The UI then uses
GET /saml/error?id=<id> to extract the error message from ns_server.

Another reason to not pass the error message in URL, is
because the following attack is possible then:
An attacker sends the link like this to the victim:

http://myserver.com:8091/ui/index.html?error="please click this
malicious link"

The user thinks it is safe because it comes from the company
they trust and clicks the malicious link.

Change-Id: Ie07201240659e41a83b312fcb4402c92ad4f7815
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195607
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>