MB-57096: Go graceful shutdown signals
Uses TERM signal to stop goport programs, wait 10 minutes and then kill
the process, if it hasn't ended on it's own already. The goal for
services, then, would be to make sure to handle this signal, and exit
cleanly after it's received.

Introduced a new setting testingGracefulShutdown that allows goport to
use this new feature for stopping services gracefully.

It appears as though some services already support this, in some
fashion (query, for example), but the remaining services will also need
to add custom handlers for their specific case(s).

To set this use the environment variable: TESTING_GRACEFUL_SHUTDOWN=true

Change-Id: I26dea248b5a367df92ebf45c48084734532bc7c3
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195663
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>