MB-57062: Add cm_rebalance_stuck metric matching the alert
In order for control plane to take advantage of the alert, we are
exposing the state of the alert as a metric. The original intention
would have been for us to find a way of getting alerts to the control
plane (such as with a metric), but this was dropped from Trinity, so we
are instead exposing this metric as it requires no infrastructure
changes for control plane to convert this into an alert on their end.

The metric will be 1 when the alert fired last time it was checked and
0 when it didn't. As the metric is aligned with the alert, to configure
its sensitivity, you must modify the threshold for the alert like so:

POST localhost:9000/settings/alerts/limits
"stuckRebalanceThresholdKV=36000"

(The threshold is given in seconds)

Change-Id: I80ce7ecb05d19c642af7654a3615726bfc031b93
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195743
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>