MB-57991 Make sure the orchestrator is not starting ...
... new processes if it is leaving the cluster

There are times when the orchestrator is ejected from the cluster.
This may take some time to be completed. During this process, Other
nodes may still find this node to be the orchestrator and may ask
this node to do something like deleting a bucket which will fail.

Here we have added another state to the orchestrator. This new state
(ejecting) indicates that the orchestrator will not accept any new
requests because it is in the process of leaving the cluster.
In this case we will wait for the old orchestartor to die and the
new one to start and retry the operation (for couple of times).

This change will add the new state only when the cluster is trinity.

Change-Id: Ida7fbec27f78d99bdf908d549d23e985b3d04fdb
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195755
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Tested-by: Amirbahador Biglarbegian <amir.biglarbegian@couchbase.com>