MB-57366: Add kv resident ratio growth reaction
Data growth over time will use more and more resources, causing the
buckets to drop their resident ratio. We need to prevent buckets
getting below a lower limit (by default 10% RR for couchstore and 1% RR
for magma) to prevent issues in the cluster.

We enforce this with a check every 20s on each node, which we then
aggregate across all nodes to determine the RR status for each bucket.
If a bucket has any node with RR% below its limit, then we tell
memcached to disable data ingress for that bucket, on all nodes with
the bucket, including the reason for ingress being disabled so that it
can provide a meaningful error code to clients.

The guard rail is enabled by default for the provisioned config profile
(with the above defaults), but can be configured with a request such as
the following:

POST "/settings/resourceManagement/bucket/residentRatio"
{
    "enabled": "true",
    "couchstoreMinimum": 5,
    "magmaMinimum": 0.5
}

The current settings can also be fetched, for example:

GET "/settings/resourceManagement/bucket/residentRatio"

{
    "enabled": "true",
    "couchstoreMinimum": 10,
    "magmaMinimum": 1
}

All resource guard rails can be configured at once using the base
endpoint "/settings/resourceManagement", for instance:

POST "/settings/resourceManagement"
{
    "bucket": {
        "residentRatio": {
  	    "enabled": True,
            "couchstoreMinimum": 5,
            "magmaMinimum": 0.5
        }, ...
    }, ...
}

Change-Id: I54d1af8070e8545dd34374d3547c9034af151d2b
Reviewed-on: https://review.couchbase.org/c/ns_server/+/193772
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>