MB-58347: Verify warmup state of memcached when getting status
Recently a change was made to ns_memcached to add a status function that
returned the status of the bucket. That function would directly return
the status in the case in which it was warmed. That causes us to fail to
detect that memcached is down for a much longer period when we find that
the DCP statuses are inactive in kv_monitor. We should continue to check
the warmup status of memcached in the case in which the status is
warmed

Change-Id: I08a438bcff7aa3d5bcdbbc016985db50443d2039
Reviewed-on: https://review.couchbase.org/c/ns_server/+/195920
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>