MB-57955: Add enableCrossClusterVersioning param
This is a new bucket config parameter that can only be enabled on
bucket update. It will be tied to a synchronous retrieval of max_cas
values for all the vbuckets of the bucket, so we don't allow this to
be set to true on bucket create.

Change-Id: I45dcb1dae24510dabdd6643afe4797edc6d4ba46
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196159
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>