MB-58410: Remove use of timer:apply_all in supervisor2
timer:apply_all() does so via spawn() which may fail if we hit some
system_limit. If the spawn() in timer:apply_all() fails then the timer
is lost forever; it is never run, and the caller is never notified in
any way. This can cause supervisor2 to fail to restart a child
indefinitely, without bubbling up any error.

This change casts the message immediately instead of sending the cast
after the timer returns. This prevents us from hitting the
timer:apply_all() issue while we deal with supervisor2 in a more
long term way.

This change is an application of this erlang/otp patch:
https://github.com/erlang/otp/commit/44d5789619c70863af918ee5e2bdd7d411e68e80

Change-Id: Ifab74f657143096695184d9d9eb43b001e76140f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196306
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>