MB-58416, MB-55212 No longer infer collection properties
We no longer infer collection properties in order to not include them in
the manifest in order to save space. Inferring them lead to
confusion/bugs in the past. As an example we used to infer "maxTTL=0"
meant to not include maxTTL in the manifest. However the meaning of
maxTTL=0 is different than the absence of maxTTL. The former means to
disable TTL while the later means to "use the bucket's maxTTL".

Change-Id: I8e7a8ec258bb027fcae12be831b814c0d9ac3f28
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196676
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>