MB-58365: Handle storage mode migration for guardrail limits
When a bucket storage mode migration is in progress, we will use the
more restrictive limit, for both directions of migration.

For the following examples, assume that the couchstore limit is more
restrictive than the magma limit.

For couchstore -> magma, while we still have couchstore nodes, since
the data is expected to be distributed evenly across the nodes, it
makes sense to continue to apply the couchstore restriction, even on
the magma nodes. As soon as the final couchstore node has left the
cluster, the bucket will be considered as a magma bucket, with the
new minimums.

For magma -> couchstore, it makes sense to immediately treat the bucket
as a couchstore bucket for resource limits, since that puts no nodes in
a potentially unsafe state, and ensures that the new nodes don't
immediately cross the couchstore resource limit as soon as they are
added (not accounting for data growth, which we would still be
correctly handling).

Change-Id: I8b51c0162500a9b61c9106387bd74c9dce29c254
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196555
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>