Use correct default for collection's maxTTL
When validating the arguments for creating a collection we were using
the incorrect default (0) for maxTTL. The default maxTTL for a
collection is "omitted" (not specified) which specifies the bucket's
maxTTL should be used.

Also, TTL is only available on Enterprise Edition so don't allow it to
be specified during collection creation on Community Edition.

Change-Id: I4fad7f8d7613778e2418842c856f53fb2b8270d2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196569
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>