MB-58365: Don't ignore name when getting stats for guardrails
The option 'ignoring(name)' is added by promQL:op/2, but this means
that sometimes not all requested stats are fetched. If the query for
two of the stats fetched has the same set of labels, then the only
label differentiating them is the name, so adding ignoring(name) means
that second the stat is lost.

For context, ignoring(name) is usually wanted for other ops as it
ensures that operations over different metrics can ignore the name and
use both metrics in the calculation. We wouldn't need to do this if we
didn't add the name label as a copy of __name__, but without the name
label we wouldn't have the behaviour where 'Metric1 or Metric2' always
gets both metrics when they have different names.

Change-Id: I5674b55b8cf21bcdd08b1139ffc6e6b47828215d
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196778
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>