MB-58257: Remove persisted lease upon node leave
When a node leaves the cluster, the lease is not releveant in the
soon to be the old life of the node in the cluster, so we clean
it up now as part of node leave cleanup. Leaving it on disk will
cause it to be recovered again when the same node newly joins the
cluster, and that is incorrect.

Change-Id: Idb75a84076dae127c7ec691131d1509c3d01ecba
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196814
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>