MB-58527 Fix broken dcp stats logging
We gather dcp stats from memcached every 10 minutes and attempt to
log them. Whilst we log "logging dcp stats" and "end of logging
dcp stats" messages, we don't actually appear to log any dcp stats
anymore.

This is due to the received stats doesn't match the expected pattern
in the logging function. As the stats are so large, even if we fix
the issue, the data is truncated when logging. So probably it is not
useful at all.

After some discussion, it seems like removing the whole code for
logging these stats is the best solution.

Change-Id: I01c458699ed80fc595a8d7d177c281d8a602ee60
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196824
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>