MB-58580 Update queryNodeQuota only when trinity
Currently we try to update queryNodeQuota in the query setting manager
on a /pools/default POST. This cannot be done if the cluster compat mode
is not trinity. This case occurs when --pretend-version is specified for
cluster_run. With this change the update won't occur if the cluster
compat mode isn't trinity.

Change-Id: Ib1ceabfdad5bdac092425fae4f179018d4fc7831
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196876
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>