Add enable_auto_rebalance_settings to provisioned_profile
When provisioned_profile was created, this key was not added, as it is
assumed that config profiles should only need to include keys that they
change. This key is required in the default profile however, so it
should also be in the provisioned profile, as we do not wish to disable
auto rebalance settings in provisioned clusters.

To avoid mistakes like this in future, we should at some point replace
this key with a new one with the inverse behaviour, so that the key can
be removed from the default_profile and provisioned_profile. This is
detailed in MB-58579.

Change-Id: I00cb9c62ce94dff1ec9dc02006a3d8b99ef4bd42
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196925
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Bryan McCoid <bryan.mccoid@couchbase.com>