MB-58506: add onunload handler
Browser implements something we can call the “bfcache”
(back-forward cache). This cache stores the
current state of the DOM in memory. When a
page is re-visited via the back button, no
requests are made. The DOM itself is served
straight up from cache.

window.unload does nothing but break the bfcache.

source:
https://madhatted.com/2013/6/16/you-do-not-understand-browser-history

Change-Id: I2a6c49177eaa24730404924014fc605f6be30d24
Reviewed-on: https://review.couchbase.org/c/ns_server/+/196970
Reviewed-by: Raluca Lupu <raluca.lupu@couchbase.com>
Tested-by: Pavel Blagodov <stochmail@gmail.com>