MB-58541: Add enableShardAffinity indexer setting
Adds the setting "enableShardAffinity" with various defaults depending
on configuration profile that is enabled. Additionally, there exists a
configuration key that indexing will place into metakv iff they are
attempting to use the shardAffinity/file-based-rebalance before a full
upgrade (of all nodes) to 7.6. Then, during upgrade to 7.6, this setting
is read, and iff it exists, we use that as the default value. If this
key is not present, the default will be configured based on the current
profile as previously mentioned.

Change-Id: I411adc1c1d0ba02f9ed2f786c826ae8cfd344ab2
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197705
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Steve Watanabe <steve.watanabe@couchbase.com>