MB-58541: Add enableShardAffinity indexer setting
Adds the setting "enableShardAffinity" with various defaults depending
on configuration profile that is enabled. Because of this it is tied to
a config_profile setting with an inverse name called
disableShardAffinity s/t it can be enabled in the default profile and
not have to also have it disabled in the other profiles.

Change-Id: Ief76596fa5f3ba7a3603f07d37181681e74271f1
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197047
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>