MB-58636: Set janitor_agent:wait_for_memcached Parent pid correctly
We currently use misc:executing_on_new_process here to allow us to use
timer to kill child processes should they take too long. This potential
process killing requires us to trap exits in the new process body and
process the EXIT messages manually to avoid taking down the janitor
should we time out here. The code anticipates that the parent process
may be killed and has a receive clause that causes it to exit itself,
however, the Parent pid passed down to match this EXIT message is
incorrect due to the use of async in misc:executing_on_new_process. The
Parent pid is set to self() outside of the misc:executing_on_new_process
scope. misc:executing_on_new_process uses asyncs though which creates a
controller process linked to the outside scope, and an executor process
linked to the controller. The executor is the process in the body of
misc:executing_on_new_process and is currently waiting for a message
from the outer scope process, to which it is not linked. When the EXIT
signal comes down to cancel the janitor run the outer scope process
sends it to the linked controller process which sends it to the executor
process which then ignores it. Change the parent pid in the executor
process to the result of async:get_identity() which returns the
controller pid.

In a future change we should investigate rewriting this code. At a
glance it lends itself quite well to misc:parallel_map, apart from the
way that it deals with warmup messages (this still hits the timeout but
may send/process a message to indicate it's actually warming up). As
such, it's unfortunately not a trivial rewrite so deferring that.

Change-Id: Ic99a7348ef4d3b9900ba366a2adc4fb06de317f9
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197185
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>