Fix rank handling in mixed-version clusters
Make it s/t we can still use our --pretend-version feature to test mixed
version clusters by modifying how we handle the presence of the "rank"
bucket parameter. Previously it would complaign joining nodes in
mixed-mode regardless of if it was explicitly set or not. Now, if it's
not explicitly set, we don't assume a default value and instead just
ignore it.

Change-Id: I5c1a3da3e97fff537ce3a67a2f8b7040828e85fd
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197207
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Neelima Premsankar <neelima.premsankar@couchbase.com>
Tested-by: Bryan McCoid <bryan.mccoid@couchbase.com>