[cluster_tests] Poll nodes after rebalance rather than assert
For ServicelessNodeTests we test removing the orchestrator node, which
was sometimes failing on the assert because the node hadn't completely
been ejected yet. This was because we eject the orchestrator node
after updating the rebalance status.

By polling for the nodes list, we give time for the orchestrator to be
ejected and /pools/default becoming consistent with connected_nodes.

Change-Id: Ice5b99eb14e4d434d11069adc2c3c59771796bed
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197329
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>