MB-58751: Don't reset cm_rebalance_stuck to 0 between checks
When the alert check occurs before it is due to check rebalance
progress again, it was setting the cm_rebalance_stuck guage back to 0,
until it next checked the rebalance progress. We should not do this, as
the guage should stay at 1 until the rebalance makes progress or stops.

Change-Id: If5f68e8cdd2be40b75fe736e3084c20b2b3675f8
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197460
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>