MB-58762: Refactor auto_failover_cfg state manipulation into common fun
When we enable auto_failover we set up the process state in a certain
order to ensure that we correctly set the down_threshold (number of
ticks) of auto_failover_logic. This wasn't being done when we enabled
auto_failover initially, but would later be updated via a cfg change
which was confusing to read in logs. Pull out the correct code from the
auto_failover settings modification function, and use that in both
places.

Change-Id: I51cb1964a2e5ac28f9008fda2c1c4dc26f389842
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197553
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>