MB-58744: Move all testsets into a dedicated directory
It is confusing having the test modules in the same directory as
run.py, README, requirements.txt, etc.
They should be in a dedicated testsets directory, so that when we are
looking for a testset, we only need to consider testsets, and vice
versa.

We should probably have done this from the start, but better late than
never.

Change-Id: I549fc987463d03288f69f8d6e305b5999fff1e15
Reviewed-on: https://review.couchbase.org/c/ns_server/+/192716
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Peter Searby <peter.searby@couchbase.com>