Collect more data from memcached in collect_info
Add histograms for connection scheduling (individual and
the aggregate).

Move bucket_details from the "per bucket" command to the
global pool (as it returns the entire array of all buckets
and it is no point of running it multiple times)

Add histograms on a per-bucket base for the amount of time
spent doing JSON validation and SNAPPY inflation.

Change-Id: I2b27f3e0d1fca6997045d765e2382b32c4803122
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197600
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>