MB-58946: Use Service enum throughout cluster tests
This reimplements https://review.couchbase.org/c/ns_server/+/197619
which was reverted as the Service enum wasn't being used in
HardResetTests, causing the test to fail when getting requirements.

In addition to the changes in the original patch, this change also
makes use of the Service enum for Cluster.add_node and
Cluster.do_join_cluster, so that the same service list can be used for
those as well as the ClusterRequirements. At the same time, most of the
original code was refactored to allow easier use of the Service enum
in the additional places.

Original commit message:

When we inevitably try to use "query" instead of "n1ql", we should
receive a clear error, instead of the cluster just failing to connect
(See MB-58805 for the lack of error message when we fail to connect)

To enforce this, a Service enum is added, which should be used instead
of strings.

Change-Id: I6dbb9d40014a6c3b4c65e4ee50495ca7ff258850
Reviewed-on: https://review.couchbase.org/c/ns_server/+/198206
Tested-by: Peter Searby <peter.searby@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>