cluster_connect: Avoid confusion of "param": args.dont_param
For replica_index and do_rebalance, we have args
"--dont-index-replicas" and "--dont-rebalance" which disable the
respective features. In the code we have

"replica_index": args.dont_index_replicas

which looks wrong. The code is technically correct, as it uses
store_false to ensure that args.dont_index_replicas is actually
`not dont-index-replicas`.

It is much simpler to just use 'store_true' and add a `not` where it
would be expected.

Change-Id: Ide3ff27ec8d856697bde09922d18c741c642fa4f
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197640
Tested-by: Peter Searby <peter.searby@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>