MB-58330: Add safeguards to prevent config corruption
This change adds validation to the chronicle transaction for node
removal such that it will reject the update if it finds any of the to
be removed nodes to be active kv nodes that are present in the server
list of any of the configured buckets. It will also reject it if it
finds any other services active on the node.

This also fixes the delta recovery config update such that it rejects
the config update for a node being added back in, if the to be
recovered node being added back in is not part of nodes_wanted.

Both checks are important because if we allow such updates in these
scenarios, they result in guaranteed config corruption.

Change-Id: Ifb63e6481ff7535b64921bd6994142eadf9dc278
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197644
Tested-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Artem Stemkovski <artem@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>