MB-58697 Add {history, false} on upgrade (as applicable)
The absence of the "history" property in the manifest collection is
equivalent to {history,false}. The direction we're moving towards is
not having inferrences based on the omission of a property. So on
upgrade from a prior releases where "history" was omitted we fill it in.

Change-Id: Idd04164641c2e95fd4103340b4456d5788a049c5
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197648
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Steve Watanabe <steve.watanabe@couchbase.com>