MB-58846: Handle unexpected normal exit in run_external_tool
A goport child process can exit normally without sending us an exit
status in some yet to be determined case. This is currently handled
poorly by run_external_tool which does not trap exits and causes
erlang to eat the {'EXIT', goport, normal} message.

This was observed in a live cluster in which we saw that the async
executor was stuck in the receive block waiting for an expected response
from the goport process. The async executor was ONLY linked to the async
controller, it was no longer linked to the goport process, as the goport
process had exited, thus, an expected emssage was never going to be
received by the async executor process.

Trap exits here, to have erlang send us the message, so that we can
take down the process when it isn't behaving as expected.

Change-Id: Icb1ae3e73e55b05c0f369a75d6b209dc3aa47399
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197773
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>