MB-58693: Add expiry and preserveTTL to .../docs endpoint
To avoid nullifying the original TTL setting when modifying a document
using the /pools/default/buckets/<bucket>/.../docs/<doc_id> we are
supporting an integer 'expiry' arg and a boolean 'preserveTTL' arg.

Example of upserting a doc, disabling expiry:

POST /pools/default/buckets/<bucket>/docs/<doc>
-d expiry=0

Example of upserting a doc, setting expiry to 100s:

POST /pools/default/buckets/<bucket>/docs/<doc>
-d expiry=100

If expiry is not specified, the behaviour is unchanged, that is, when
we upsert the doc, expiry is disabled. This can be avoided with the
preverseTTL flag as follows.

Example of upserting a doc without resetting expiry:

POST /pools/default/buckets/<bucket>/docs/<doc>
-d value=<new_value>
-d preserveTTL=true

Example of updating a document, specifying an expiry, but ignoring it
and preserving the existing expiry:

POST /pools/default/buckets/<bucket>/docs/<doc>
-d expiry=100
-d preserveTTL=true

If preserveTTL, is not specified, or is false, then expiry will always
implicitly be set to 0.

Change-Id: I8b8390029f71668dff82bdc7274e359caa2e6e4c
Reviewed-on: https://review.couchbase.org/c/ns_server/+/197848
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>